Remove Y up as special case
Added root transform nodes for RigidBodies and Skeletons to allow for a Z to Y rotation, this is the only behavior now in Y up mode (it was doing stuff for the old QAvatar before).  Made the Z to Y conversion a rotation instead of swapping translation values.

Added support for new XML solver definitions (multiple solvers and changed hierarchy).